# server/src/middleware/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Faq": {
				"properties": {
					"id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"question",
					"answer",
					"categoryId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetFaqResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Faq"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_createFaqRequest_": {
				"properties": {
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"updateFaqRequest": {
				"$ref": "#/components/schemas/Partial_createFaqRequest_"
			},
...
			"ExternalResource": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"url",
					"categoryId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllExternalResourceResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ExternalResource"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetExternalResourceResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ExternalResource"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ExternalResource.Exclude_keyofExternalResource.id__": {
				"properties": {
					"categoryId": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"categoryId",
					"url",
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ExternalResource.id_": {
				"$ref": "#/components/schemas/Pick_ExternalResource.Exclude_keyofExternalResource.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"createExternalResourceRequest": {
				"$ref": "#/components/schemas/Omit_ExternalResource.id_"
			},
			"Partial_createExternalResourceRequest_": {
				"properties": {
					"categoryId": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"updateExternalResourceRequest": {
				"$ref": "#/components/schemas/Partial_createExternalResourceRequest_"
			},
			"ExternalResourceCategory": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllExternalResourceCategoryResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ExternalResourceCategory"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetExternalResourceCategoryResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ExternalResourceCategory"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ExternalResourceCategory.Exclude_keyofExternalResourceCategory.id__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ExternalResourceCategory.id_": {
				"$ref": "#/components/schemas/Pick_ExternalResourceCategory.Exclude_keyofExternalResourceCategory.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"createExternalResourceCategoryRequest": {
				"$ref": "#/components/schemas/Omit_ExternalResourceCategory.id_"
			},
			"Partial_createExternalResourceCategoryRequest_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"updateExternalResourceCategoryRequest": {
				"$ref": "#/components/schemas/Partial_createExternalResourceCategoryRequest_"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/faq/{id}": {
			"get": {
				"operationId": "GetFaqCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetFaqResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateFaq",
				"responses": {
					"200": {
						"description": "Successfully updated FAQ",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetFaqResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateFaqRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFaq",
				"responses": {
					"204": {
						"description": "Successfully deleted FAQ"
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/external-resources": {
			"get": {
				"operationId": "GetAllExternalResources",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAllExternalResourceResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateExternalResource",
				"responses": {
					"201": {
						"description": "Successfully created external resource",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createExternalResourceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAllExternalResources",
				"responses": {
					"204": {
						"description": "Successfully deleted all External Resources"
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/external-resources/{id}": {
			"get": {
				"operationId": "GetExternalResource",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateExternalResource",
				"responses": {
					"200": {
						"description": "Successfully updated external resource",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateExternalResourceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExternalResource",
				"responses": {
					"204": {
						"description": "Successfully deleted External Resource"
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/external-resource-category": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAllExternalResourceCategoryResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Successfully created External Resource Category",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceCategoryResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createExternalResourceCategoryRequest"
							}
						}
					}
				}
			}
		},
		"/external-resource-category/{id}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceCategoryResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Successfully updated FAQ category",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetExternalResourceCategoryResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateExternalResourceCategoryRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "Successfully deleted FAQ category"
					}
				},
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}